HDDS-12054. Move ozone debug prefix to ozone debug om prefix
